+++ apps/rig_api/lib/rig_api/controllers/channels_controller.ex
@@ -5,46 +5,11 @@ defmodule RigApi.ChannelsController do
   """
   use RigApi, :controller
   use Rig.Config, [:session_role]
-  use PhoenixSwagger
   require Logger
   alias RigAuth.Blacklist
   alias RigAuth.Jwt
   alias RigInboundGatewayWeb.Endpoint
   alias RigInboundGatewayWeb.Presence.Channel
-
-  def swagger_definitions do
-    %{
-      User:
-        swagger_schema do
-          title("tbd")
-          description("tbd")
-
-          properties do
-            userId(:string, "tbd", required: true) 
-          end
-
-          example(%{
-            userId: "SomeUserId"
-          })
-        end, 
-      Users:
-        swagger_schema do
-          title("tbd")
-          description("tbd")
-          type :array 
-          items Schema.ref(:User)
-        end
-      }
-  end
-
-  # Swagger documentation for endpoint GET /v1/users
-  swagger_path :list_channels do
-    get("/v1/users")
-    summary("tbd")
-    description("tbd")
-
-    response(200, "Ok", Schema.ref(:Users))
-  end
 
   def list_channels(conn, _params) do
     conf = config()
@@ -57,19 +22,6 @@ defmodule RigApi.ChannelsController do
     json(conn, channels)
   end
 
-  # Swagger documentation for endpoint GET /v1/users/:user/sessions
-  swagger_path :list_channel_sessions do
-    get("/v1/users/{user}/sessions")
-    summary("tbd")
-    description("tbd")
-
-    parameters do
-      user(:path, :string, "tbd", required: true)
-    end
-
-    response(200, "Ok")
-  end
-
   def list_channel_sessions(conn, %{"user" => id}) do
     sessions =
       "user:#{id}"
@@ -80,18 +32,6 @@ defmodule RigApi.ChannelsController do
     json(conn, sessions)
   end
 
-  # Swagger documentation for endpoint DELETE /v1/tokens/:jti
-  swagger_path :disconnect_channel_session do
-    delete("/v1/tokens/{jti}")
-    summary("tbd")
-    description("tbd")
-
-    parameters do
-      jti(:path, :string, "tbd", required: true)
-    end
-
-    response(200, "Ok")
-  end
   def disconnect_channel_session(conn, %{"jti" => jti}) do
     Blacklist.add_jti(Blacklist, jti, jwt_expiry(conn))
     Endpoint.broadcast(jti, "disconnect", %{})